Fix schemas involving `Maybe`
See the previous commit for an explanation of the problem.

This change also rules out `ToSchema (Maybe (Maybe _))` for correctness.

---

The schema for `Maybe a` is prefixed with `Maybe_` to distinguish it from
the schema for `a`. Without this, `declareSchemaRef` will generate the
same reference for an `a` or a `Maybe a`. When an API uses an `a` and a
`Maybe a`, only one of the two schemas will make it into the
`components.schemas` section (because of the name conflict). This can
lead to a schema that says it produces non-nullable values, but may
actually produce nulls.

This problem could also be solved by changing `declareSchemaRef` to
strip the `nullable` property from the declared schema and return the
nullability value alongside the `Referenced Schema`. This would increase
sharing (you wouldn't have near duplicate schemas `X` and `Maybe_X`), with
the cost of forcing downstream libraries (such as `servant-openapi3`) to
explicitly handle that nullability information. I chose the solution that's
compatible with current downstream code.